code-agent/requirements.txt
crewai
crewai-tools
python-dotenv
plotly
pandas
llama-index
llama-index-tools-yahoo-finance 